turn large dashboard file into a bunch of smaller files with their own purposes
