+++ client/src/components/Calendar/CalendarNav.css
@@ -1,0 +1,5 @@
+.Calendar_nav {
+  display: flex;
+  margin: 1rem;
+  font-size: 85%;
+}
