+++ client/src/components/Calendar/Weekdays.css
@@ -1,0 +1,14 @@
+.Weekdays {
+  display: flex;
+  flex: 1;
+  margin: 2px;
+}
+
+.Weekdays > * {
+  text-transform: uppercase;
+  flex: 1;
+  text-align: center;
+  font-weight: bold;
+  color: var(--black);
+  font-size: 85%;
+}
